refactor: 폴더구조 변경 및 리스트 더 불러오기 기능 구현
