+++ frontend/src/components/@shared/Card/Card.styles.ts
