+++ frontend/src/components/@shared/Card/Card.tsx
