+++ frontend/src/components/Banner/Banner.styles.ts
