+++ frontend/src/components/Card/ReviewCard.tsx
@@ -3,10 +3,10 @@ import { useParams } from 'react-router-dom';
 import { COLOR } from 'src/constants';
 import UserContext from 'src/contexts/UserContext';
 import nicknameGenerator from 'src/utils/createNickname';
+import Card from '../@shared/Card/Card';
 import HumanIcon from '../Icons/human';
 import { modalContext } from '../Modal/ModalProvider';
 import ReviewEditForm from '../Review/ReviewEditForm';
-import Card from './Card';
 import { Header, ReviewerInfo, Content, ShowMoreButton } from './ReviewCard.styles';
 
 interface Props {
