+++ frontend/src/components/Item/CardItem.styles.ts
