+++ frontend/src/components/Item/ListItem.styles.ts
