+++ frontend/src/components/List/CardList.styles.ts
