+++ frontend/src/components/Review/ReviewCreateForm.tsx
@@ -4,7 +4,7 @@ import { useParams, useHistory } from 'react-router-dom';
 import API from 'src/apis/requests';
 import { COLOR, MESSAGE, ERROR_MESSAGE, REVIEW, PATH } from 'src/constants';
 import UserContext from 'src/contexts/UserContext';
-import Card from '../Card/Card';
+import Card from '../@shared/Card/Card';
 import { Form } from './ReviewCreateForm.styles';
 
 const LOGGED_IN_PLACEHOLDER =
